+++ CHANGELOG.md
@@ -40,3 +40,16 @@
 ## MISSING
 
 - Tests
+- Avatar
+- Calendar
+- Carousel
+- Checkbox
+- Combobox
+- Data Table
+- Dialog
+- Sheet
+- Sidebar
+- Skeleton
+- Sonner
+- Switch
+- TextArea