+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 
 ### Features
 
+* :sparkles: add new Badge component ([2d85dec](https://github.com/Fixed-My-City/frontend/commit/2d85dec172e5165a87cffa775cd358242b4e6ff9))
+* :sparkles: add new Button component and Shadcn ([63905ef](https://github.com/Fixed-My-City/frontend/commit/63905eff28f9b8ea9901d861c8b122f32272fda8))
+* :sparkles: add new Card component ([79f6d13](https://github.com/Fixed-My-City/frontend/commit/79f6d13a4d51b9ad511cc84d8ecee45bab96157d))
+* :sparkles: add new Input component and Shadcn ([0cff73b](https://github.com/Fixed-My-City/frontend/commit/0cff73bc52486ef73a8e4d98e28e5a3bd5383bbb))
+* :sparkles: add new Label component ([7660fa1](https://github.com/Fixed-My-City/frontend/commit/7660fa1b87d67c57f1df0879bd7718a1bf8a44bb))
+* :sparkles: add new Select component and Shadcn ([b25dbc9](https://github.com/Fixed-My-City/frontend/commit/b25dbc9622670b1adff9b01ac587de09ae569dfb))
+* :sparkles: add next auth ([64d70b0](https://github.com/Fixed-My-City/frontend/commit/64d70b0b47b080bcf03acf13093678ce1eed5de3))
+* :sparkles: add next auth ([e7d2fc2](https://github.com/Fixed-My-City/frontend/commit/e7d2fc28c310d1fed829a15a1392c3ad8d23a959))
+* :sparkles: first commit ([cfc5b74](https://github.com/Fixed-My-City/frontend/commit/cfc5b74df0b696a75c0396525cfde50c194b0e15))
+
+# 1.0.0 (2025-01-07)
+
+
+### Features
+
 * :sparkles: add new Button component and Shadcn ([63905ef](https://github.com/Fixed-My-City/frontend/commit/63905eff28f9b8ea9901d861c8b122f32272fda8))
 * :sparkles: add new Input component and Shadcn ([0cff73b](https://github.com/Fixed-My-City/frontend/commit/0cff73bc52486ef73a8e4d98e28e5a3bd5383bbb))
 * :sparkles: add new Select component and Shadcn ([b25dbc9](https://github.com/Fixed-My-City/frontend/commit/b25dbc9622670b1adff9b01ac587de09ae569dfb))